Move kernels to separate directory/files
